Rename course to new name
821e83aa28 / Change title of IBA course
