+++ pages/courses/image-based-abuse-and-rebuilding-ourselves/[sessionSlug].tsx
